make util/index.js not pull in all of util/io
