+++ packages/jbrowse-web/src/util/index.js
@@ -1,23 +1,11 @@
 import fromEntries from 'object.fromentries'
 
-import { openUrl } from './io/rangeFetcher'
-import LocalFile from './io/localFile'
-
 if (!Object.fromEntries) {
   fromEntries.shim()
 }
 
 export function assembleLocString({ assemblyName, refName, start, end }) {
   return `${assemblyName}:${refName}:${start + 1}-${end}`
-}
-
-export function openLocation(location) {
-  if (!location) throw new Error(`must provide a location to openLocation`)
-  if (location.uri) return openUrl(location.uri)
-  if (location.path) {
-    return new LocalFile(location.path)
-  }
-  throw new Error('invalid fileLocation')
 }
 
 export function clamp(val, min, max) {
